\docType{data}
\name{format.ASPT}
\alias{format.ASPT}
\title{A formatting scheme following ASPT (American Society of Plant Taxonomists) conventions for an exsiccatae record.}
\format{A dataframe with 9 entries and three variables
\itemize{
 \item{"field"} {A column name present in the exsiccatae table.}
 \item{"style"} {The text style applied to that element. One of: none, bold, italics, capitals,
 underline or underscore, uppercase, and () or parentheses}
 \item{"sept"}  {A seperator following the element. One of '', ' ', ', ', '. ', ': ' and ';' }
 \item{"comments"} {}
}}
\source{
  \url{http://www.aspt.net/publications/sysbot/checklist_systbot.php}
}
\description{
  A formatting scheme following ASPT (American Society of
  Plant Taxonomists) conventions for an exsiccatae record.
}
\author{
  R. Simon
}
\seealso{
  Other formats: \code{\link{format.NYBG}},
  \code{\link{format.PK}}, \code{\link{format.SBMG}}
}
\keyword{datasets}

