\docType{data}
\name{format.SBMG}
\alias{format.SBMG}
\title{A formatting scheme following SBMG (Systematic Botany Monographs) conventions for an exsiccatae record.}
\format{A dataframe with 9 entries and three variables
\itemize{
 \item{"field"} {A column name present in the exsiccatae table.}
 \item{"style"} {The text style applied to that element. One of: none, bold, italics, capitals,
 underline or underscore, uppercase, and () or parentheses}
 \item{"sept"}  {A seperator following the element. One of '', ' ', ', ', '. ', ': ' and ';' }
 \item{"comments"} {}
}}
\description{
  A formatting scheme following SBMG (Systematic Botany
  Monographs) conventions for an exsiccatae record.
}
\author{
  R. Simon
}
\seealso{
  Other formats: \code{\link{format.ASPT}},
  \code{\link{format.NYBG}}, \code{\link{format.PK}}
}
\keyword{datasets}

