\name{index.citations}
\alias{index.citations}
\title{Creates an index of specimen citations}
\usage{
  index.citations(atable, format = format.SBMG)
}
\arguments{
  \item{atable}{a table with exsiccatae records}

  \item{format}{a data frame containing formatting options}
}
\value{
  a string in markdown format
}
\description{
  Creates an index of specimen citations
}
\author{
  Reinhard Simon
}
\seealso{
  Other main: \code{\link{exsic}},
  \code{\link{index.collections}},
  \code{\link{index.countries}},
  \code{\link{index.species}},
  \code{\link{index.species.short}},
  \code{\link{prepare.table}}, \code{\link{read.exsic}},
  \code{\link{section.exsic}}, \code{\link{write.exsic}}
}

