% Generated by roxygen2 (4.0.2): do not edit by hand
\name{index.countries}
\alias{index.countries}
\title{Creates a simple index of countries.}
\usage{
index.countries(atable = NULL)
}
\arguments{
\item{atable}{a table with exsiccatae records}
}
\value{
a string in markdown format
}
\description{
Creates a simple index of countries.
}
\author{
Reinhard Simon
}
\seealso{
Other main: \code{\link{exsic}};
  \code{\link{index.citations}};
  \code{\link{index.collections}};
  \code{\link{index.species.short}};
  \code{\link{index.species}}; \code{\link{prepare.table}};
  \code{\link{read.exsic}}; \code{\link{section.exsic}};
  \code{\link{write.exsic}}
}

