% Generated by roxygen2 (4.0.2): do not edit by hand
\name{index.species}
\alias{index.species}
\title{Creates a simple index of species.}
\usage{
index.species(atable = NULL)
}
\arguments{
\item{atable}{a table with exsiccatae records}
}
\value{
a string in markdown format
}
\description{
Creates a simple index of species.
}
\author{
Reinhard Simon
}
\seealso{
Other main: \code{\link{exsic}};
  \code{\link{index.citations}};
  \code{\link{index.collections}};
  \code{\link{index.countries}};
  \code{\link{index.species.short}};
  \code{\link{prepare.table}}; \code{\link{read.exsic}};
  \code{\link{section.exsic}}; \code{\link{write.exsic}}
}

