\name{index.species.short}
\alias{index.species.short}
\title{A condensed list of species (listed within a line).}
\usage{
  index.species.short(atable = NULL)
}
\arguments{
  \item{atable}{a table with exsiccatae records}
}
\value{
  a string in markdown format
}
\description{
  Should only be used within the template file.
}
\author{
  Reinhard Simon
}
\seealso{
  Other main: \code{\link{exsic}},
  \code{\link{index.citations}},
  \code{\link{index.collections}},
  \code{\link{index.countries}},
  \code{\link{index.species}}, \code{\link{prepare.table}},
  \code{\link{read.exsic}}, \code{\link{section.exsic}},
  \code{\link{write.exsic}}
}

