\name{is.exsic}
\alias{is.exsic}
\title{Checks table format}
\usage{
  is.exsic(obj)
}
\arguments{
  \item{obj}{an object}
}
\description{
  Checks if the table is in the expected format by the
  principal exsic function.
}
\details{
  The format should be created either using reading a table
  via read.exsic or prepare.table.
}
\author{
  Reinhard Simon
}
\seealso{
  Other helper: \code{\link{coll.cite}},
  \code{\link{is.format}}, \code{\link{is.sortfilter}},
  \code{\link{strip.final.dot}}
}

