\name{is.format}
\alias{is.format}
\title{Checks if the data frame conforms to the expectations to a formatting configuration
data frame.}
\usage{
  is.format(fmt)
}
\arguments{
  \item{fmt}{a special data frame}
}
\value{
  boolean TRUE if ok
}
\description{
  Checks if the data frame conforms to the expectations to
  a formatting configuration data frame.
}
\author{
  Reinhard Simon
}
\seealso{
  format.SBMG, format.NYBG

  Other helper: \code{\link{coll.cite}},
  \code{\link{is.exsic}}, \code{\link{is.sortfilter}},
  \code{\link{strip.final.dot}}
}

