% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exsicnew.R
\name{section.exsic}
\alias{section.exsic}
\title{Creates a section header}
\usage{
section.exsic(title = "My title")
}
\arguments{
\item{title}{a section title}
}
\value{
a string in markdown format
}
\description{
A simple wrapper to return a section header formatted for inclusion in the final document.
}
\author{
Reinhard Simon
}
\seealso{
Other main: \code{\link{exsic}};
  \code{\link{index.citations}};
  \code{\link{index.collections}};
  \code{\link{index.countries}};
  \code{\link{index.species.short}};
  \code{\link{index.species}}; \code{\link{prepare.table}};
  \code{\link{read.exsic}}; \code{\link{write.exsic}}
}

