\name{write.exsic}
\alias{write.exsic}
\title{Writes the exsiccatae text}
\usage{
  write.exsic(text, file)
}
\arguments{
  \item{text}{a text in markdown format}

  \item{file}{a file name}
}
\description{
  Writes the text string into a file formatted as HTML.
}
\author{
  Reinhard Simon
}
\seealso{
  Other main: \code{\link{exsic}},
  \code{\link{index.citations}},
  \code{\link{index.collections}},
  \code{\link{index.countries}},
  \code{\link{index.species}},
  \code{\link{index.species.short}},
  \code{\link{prepare.table}}, \code{\link{read.exsic}},
  \code{\link{section.exsic}}
}

